requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.1
psycopg2==2.7.7
Pygments==2.3.1
python-decouple==3.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
SQLAlchemy==1.2.18
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
